Issue 1309: Time zone of feedback session could not support minute offset Update Issue 1309
--HG--
branch : release


Former-commit-id: 8a7b8433426af84167cdaa8b9ac69f677a1d5380
Former-commit-id: 286cedfd0b8bbaf1d7ecd02c4677a193b2651ec8
Former-commit-id: 2b461589f7923089140e7e172f8266104e5f13cb

+++ src/main/java/teammates/common/datatransfer/FeedbackSessionAttributes.java
@@ -30,7 +30,7 @@ public class FeedbackSessionAttributes extends EntityAttributes implements Sessi
 	public Date endTime;
 	public Date sessionVisibleFromTime;
 	public Date resultsVisibleFromTime;
-	public int timeZone;
+	public double timeZone;
 	public int gracePeriod;
 	public FeedbackSessionType feedbackSessionType;
 	public boolean sentOpenEmail;
